requirements.txt
streamlit 
pytesseract 
pymupdf 
Pillow
